updated ui and profile page

+++ react/src/components/Logout.jsx
@@ -1,27 +1,43 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import { BiPowerOff } from "react-icons/bi";
+// import { BiPowerOff } from "react-icons/bi";
+import { IoLogOut } from "react-icons/io5";
+
 import styled from "styled-components";
-import axios from "axios";
-import { logoutRoute } from "../utils/APIRoutes";
+// import axios from "axios";
+// import { logoutRoute } from "../utils/APIRoutes";
 
 export default function Logout() {
   const navigate = useNavigate();
 
-  const handleClick = async () => {
+  // const handleClick = async () => {
 //     const id = await JSON.parse(
 //                     localStorage.getItem('baat-cheet-user')
 //                 )._id;
 //     const data = await axios.get(`${logoutRoute}/${id}`);
 //     if (data.status === 200) {
+      // localStorage.clear();
+      // navigate("/login");
+    // }
+  // };
+
+  const handleLogout = () => {
+    const confirmed = window.confirm('Are you sure you want to log out?');
+  
+    if (confirmed) {
       localStorage.clear();
+      // console.log('User confirmed to log out');
       navigate("/login");
+    } 
+    // else {
+      // User canceled
+      // console.log('User canceled log out');
     // }
   };
 
   return (
-    <Button onClick={handleClick}>
-      <BiPowerOff />
+    <Button onClick={handleLogout}>
+      <IoLogOut />
     </Button>
   );
 }
